convert coopbot
