--- menubot/state.json
@@ -1,79 +1,0 @@
-{
-    "tick": 6450,
-    "sceneName": "CharSelect",
-    "gameState": {
-        "-1396362": {
-            "id": -1396362,
-            "type": "Seat7Button",
-            "isStatic": false,
-            "position": {
-                "x": 813.4953,
-                "y": 82.87045,
-                "z": 0,
-                "normalized": {
-                    "x": 0.994851351,
-                    "y": 0.101345122,
-                    "z": 0,
-                    "magnitude": 1,
-                    "sqrMagnitude": 1
-                },
-                "magnitude": 817.7054,
-                "sqrMagnitude": 668642.1
-            },
-            "rotation": {
-                "x": 0,
-                "y": 0,
-                "z": 0,
-                "w": 1,
-                "eulerAngles": {
-                    "x": 0,
-                    "y": 0,
-                    "z": 0,
-                    "magnitude": 0,
-                    "sqrMagnitude": 0
-                }
-            },
-            "interactable": true
-        },
-        "-1395284": {
-            "id": -1395284,
-            "type": "ReadyButton",
-            "isStatic": false,
-            "position": {
-                "x": 852.293457,
-                "y": 215.7338,
-                "z": 0,
-                "normalized": {
-                    "x": 0.9694263,
-                    "y": 0.245382637,
-                    "z": 0,
-                    "normalized": {
-                        "x": 0.969426334,
-                        "y": 0.245382652,
-                        "z": 0,
-                        "magnitude": 1,
-                        "sqrMagnitude": 1
-                    },
-                    "magnitude": 0.99999994,
-                    "sqrMagnitude": 0.99999994
-                },
-                "magnitude": 879.173035,
-                "sqrMagnitude": 772945.2
-            },
-            "rotation": {
-                "x": 0,
-                "y": 0,
-                "z": 0,
-                "w": 1,
-                "eulerAngles": {
-                    "x": 0,
-                    "y": 0,
-                    "z": 0,
-                    "magnitude": 0,
-                    "sqrMagnitude": 0
-                }
-            },
-            "interactable": true
-        }
-    }
-}
